+++ include/ui/metatileselector.h
@@ -45,13 +45,14 @@ public:
     QPoint getSelectionDimensions();
     void draw();
     bool select(uint16_t metatile);
-    bool selectFromMap(uint16_t metatileId, uint16_t collision, uint16_t elevation);
+    void selectFromMap(uint16_t metatileId, uint16_t collision, uint16_t elevation);
     void setTilesets(Tileset*, Tileset*);
     MetatileSelection getMetatileSelection();
     void setPrefabSelection(MetatileSelection selection);
     void setExternalSelection(int, int, QList<uint16_t>, QList<QPair<uint16_t, uint16_t>>);
     QPoint getMetatileIdCoordsOnWidget(uint16_t);
     void setLayout(Layout *layout);
+    bool isInternalSelection() const { return (!this->externalSelection && !this->prefabSelection); }
     Tileset *primaryTileset;
     Tileset *secondaryTileset;
 protected: